requirements.txt
pydantic==2.6.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
python-dotenv==1.0.0
pyyaml==6.0.1
gradio==4.44.1
